Support connecting to Adafruit IO MQTT
Adafruit use the API key as the MQTT password, this is 40 chars long so the EEPROM space for it is increased to 64 bytes. Also changed to use #defines to define the start/end/size of all the blocks in the EEPROM as it makes it easier to change. Obviously there is upgrade issues if you change these values so maybe a checksum or version number should be added. luckily for this change the MQTT password is the last value so no upgrade issues.