More Debug and Adafruit IO support
Adafruit IO need feeds of the format username/f/feed so if you want to group the EmonESP feeds together you need a prefix rather that a 'folder' so you can now set the topic separator so, for example, given the base topic of username/f/emonesp you can now change from username/f/emonesp/P0 to username/f/emonesp-P0

Updated the read me to include info on how to enable the ESP's trace messages, this also needs a change to use src_build_flags rather than build_flags to pass the version number. The src_build_flags are just not passed to the libraries and only passed when building anything in src.

Various bits of tidy-up that hopefully make it easier to follow the code.